Add files via upload
A basic CRUD application (Create, Read, Update, Delete) is a type of software that allows users to perform fundamental database operations. It enables the creation of new data entries, reading or retrieving data, updating existing data, and deleting records from the database. Such applications are essential in managing data-driven systems and are often used as a foundation for more complex software solutions.